+++ src/Core/Framework/DataAbstractionLayer/Indexer/EntityStock/EntityStockUpdater.php
@@ -7,6 +7,8 @@ use Shopware\Core\Checkout\Cart\Event\CheckoutOrderPlacedEvent;
 use Shopware\Core\Checkout\Cart\LineItem\LineItem;
 use Shopware\Core\Checkout\Order\Aggregate\OrderLineItem\OrderLineItemDefinition;
 use Shopware\Core\Checkout\Order\Aggregate\OrderLineItem\OrderLineItemEntity;
+use Shopware\Core\Checkout\Order\Aggregate\OrderTransaction\OrderTransactionDefinition;
+use Shopware\Core\Checkout\Order\OrderDefinition;
 use Shopware\Core\Checkout\Order\OrderEvents;
 use Shopware\Core\Checkout\Order\OrderStates;
 use Shopware\Core\Content\Product\ProductDefinition;
@@ -174,7 +176,7 @@ class EntityStockUpdater implements EventSubscriberInterface
             return;
         }
 
-        if ($event->getEntityName() !== 'order') {
+        if ($event->getEntityName() !== OrderDefinition::ENTITY_NAME) {
             return;
         }